Rename device_simulation file to device_component
