+++ src/tickit/core/components/device_component.py
